+++ uniffi_macros/src/export.rs
@@ -228,8 +228,17 @@ pub fn rewrite_self_type(item: &mut Item) {
 }
 
 /// Alter the tokens wrapped with the `[uniffi::export]` if needed
-pub fn alter_input(item: &Item) -> TokenStream {
+pub fn alter_input(item: &Item, all_args: proc_macro::TokenStream) -> TokenStream {
     match item {
+        // Like all remote types, we don't emit the duplicate definition itself, just its ffi/metadata etc
+        Item::Trait(_)
+            if matches!(
+                syn::parse::<attributes::ExportTraitArgs>(all_args),
+                Ok(args) if args.remote.is_some()
+            ) =>
+        {
+            quote! {}
+        }
         Item::Trait(item_trait) => alter_trait(item_trait),
         _ => quote! { #item },
     }